main page changed
